Make battery listener only listen to device power
On my laptop, I frequently use wireless devices, which also show up on acpi -i, so the current implementation of the watch script did not work for me. This change makes it so it only takes the lines that indicate a rate of which a % sign is at the end of the string or if 'remaining' is at the end of the string.

Not sure if this covers all use cases, but it works for me, because my wireless devices report 'rate information unavailable'.